+++ bin/tests/discount.sh
@@ -10,6 +10,7 @@ onlysnarfpy \
 -months 3 \
 -prefer-local \
 -keep \
+-show \
 -verbose -verbose -verbose \
 -browser "$1" \
 -source "$2"