+++ bin/tests/schedule.sh
@@ -8,7 +8,7 @@ onlysnarfpy \
 -bykeyword "pussycats" \
 -skip-download \
 -text "schedule testes" \
--schedule "6-7-2020:6:26" \
+-schedule "8-7-2020:22:30" \
 -debug-delay \
 -keep \
 -verbose -verbose -verbose \
@@ -22,8 +22,8 @@ onlysnarfpy \
 -bykeyword "pussycats" \
 -skip-download \
 -text "schedule testes" \
--date "6-7-2020" \
--time "6:26" \
+-date "8-7-2020" \
+-time "16:26" \
 -debug-delay \
 -keep \
 -verbose -verbose -verbose \
@@ -37,7 +37,7 @@ onlysnarfpy \
 -bykeyword "pussycats" \
 -skip-download \
 -text "schedule testes" \
--date "6-7-2020" \
+-date "8-7-2020" \
 -debug-delay \
 -keep \
 -verbose -verbose -verbose \